upgrade and migrate sveltekit
