+++ package.json
@@ -7,11 +7,9 @@
   "private": true,
   "packageManager": "pnpm@7.2.1",
   "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "svelte-kit build",
-    "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
-    "prepare": "svelte-kit sync",
+    "dev": "vite dev",
+    "build": "vite build",
+    "preview": "vite preview",
     "start": "turbo run preview",
     "format": "prettier --plugin-search-dir=. --write .",
     "format:check": "prettier --plugin-search-dir=. --list-different .",
@@ -29,7 +27,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.22.2",
-    "@sveltejs/adapter-cloudflare": "1.0.0-next.23",
+    "@sveltejs/adapter-cloudflare": "1.0.0-next.31",
     "@sveltejs/kit": "next",
     "@typescript-eslint/eslint-plugin": "^5.27.0",
     "@typescript-eslint/parser": "^5.27.0",
@@ -43,12 +41,13 @@
     "prettier": "^2.6.2",
     "prettier-plugin-svelte": "^2.7.0",
     "suncalc": "^1.9.0",
-    "svelte": "^3.44.0",
-    "svelte-check": "^2.7.1",
+    "svelte": "^3.49.0",
+    "svelte-check": "^2.8.1",
     "svelte-preprocess": "^4.10.5",
     "tslib": "^2.3.1",
     "turbo": "^1.2.16",
-    "typescript": "^4.7.2"
+    "typescript": "^4.7.2",
+    "vite": "^3.0.8"
   },
   "type": "module"
 }
