+++ src/app.d.ts
@@ -1,27 +1,15 @@
 /// <reference types="@sveltejs/kit" />
-
-import type { Theme, ClientGeolocation } from "$lib/get-sun-theme";
+/// <reference types="@sveltejs/adapter-cloudflare" />
+/* eslint-disable @typescript-eslint/no-empty-interface */
 
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
-declare global {
-  namespace App {
-    // interface Locals {}
+declare namespace App {
+  interface Locals {}
 
-    // From https://github.com/sveltejs/kit/tree/master/packages/adapter-cloudflare#environment-variables
-    interface Platform {
-      context: {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        waitUntil(promise: Promise<any>): void;
-      };
-    }
+  interface Platform {}
 
-    interface Session {
-      clientGeolocation: ClientGeolocation;
-    }
+  interface PrivateEnv {}
 
-    interface Stuff {
-      theme: Theme;
-    }
-  }
+  interface PublicEnv {}
 }
